move import attributes to attributes module
